back/requirements.txt
fastapi[all]
uvicorn
beanie
motor
pytz
PyJWT
passlib
bcrypt
python-jose
password-validator
email-validator
websockets